Expand Opera browser detection
Referenced [this page](http://my.opera.com/community/openweb/idopera/) in
order to detect versions and OSes.

OS detection wasn't added for Opera versions 8 and 9 because (a) it's a
tad more involved since the Opera version comes before the OS in the
User Agent string, and (b) Opera 10 came out in 2009. Like 4 years ago.

I added the slash in the `opera8` and `opera9` patterns because that's
what the strings in the aforementioned Web page have.

I tried something like this at first

    "opera/9.80";operagte10;B;;
    "windows";operagte10w;B;operagte10;
    "macintosh";operagte10m;B;operagte10;
    "x11";operagte10x11;B;operagte10;
    "version/10";opera10x11;C;operagte10x11;browser=Opera,version=10000,os=Linux
    "version/10";opera10w;C;operagte10w;browser=Opera,version=10000,os=Windows
    "version/10";opera10m;C;operagte10m;browser=Opera,version=10000,os=Macintosh
    "version/11";opera11x11;C;opera10x11;browser=Opera,version=11000,os=Linux
    "version/11";opera11w;C;operagte10w;browser=Opera,version=11000,os=Windows
    "version/11";opera11m;C;operagte10m;browser=Opera,version=11000,os=Macintosh
    "version/12";opera12x11;C;operagte10x11;browser=Opera,version=12000,os=Linux
    "version/12";opera12w;C;operagte10w;browser=Opera,version=12000,os=Windows
    "version/12";opera12m;C;operagte10m;browser=Opera,version=12000,os=Macintosh
    "opera/8";opera8;W;;browser=Opera,version=8000,os=
    "opera/9";opera9;W;;browser=Opera,version=9000,os=
    "opera";opera;N;;browser=Opera,version=0,os=

but I like the final version a bit more.